docs(I65__S69): consolidate code-style documentation
- Consolidate JSDoc templates: single source in documentation/jsdoc.md
- Consolidate Import Order: single source in conventions/imports.md
- Consolidate useMediaQuery example: single source in examples/hooks.md
- Simplify Decision Trees: link to conventions/use-client.md
- Consolidate TypeScript rules: detailed in typescript.md and 02-types.md
- Consolidate Constants rules: single source in 03-utils.md
- Remove duplicate directory structures: link to examples/timeline.md
- Simplify validation checklists: 3-5 critical items + link to validation.md

Reduced from 2432 to 2288 lines (144 lines, ~6% reduction)
Established clear cross-references with anchors
Maintained local context with mini-summaries
