+++ docs/folder-structure.md
@@ -1,0 +1,58 @@
+## 📂 File Naming
+
+### Next.js Pages
+```
+src/app/
+├── page.view.tsx           # Page component (separated view logic)
+├── layout.tsx              # Layout component
+└── background/
+    ├── page.view.tsx       # Route-specific page view
+    └── styles.css          # Route-specific styles
+```
+
+### Components
+```
+ComponentName.tsx           # PascalCase for components
+component-name.types.ts     # kebab-case for supporting files
+component-name.utils.ts
+component-name.constants.ts
+```
+
+## 🔄 Context API Pattern
+
+### File Structure
+```
+src/contexts/
+└── BackgroundContext.tsx    # Context + Provider in single file
+```
+
+### Naming Convention
+```tsx
+// Context name: [Feature]Context
+export const BackgroundContext = createContext<BackgroundContextType | undefined>(undefined);
+
+// Provider name: [Feature]Provider
+export function BackgroundProvider({ children }: { children: ReactNode }) {
+  // implementation
+}
+
+// Hook name: use[Feature]
+export function useBackground() {
+  const context = useContext(BackgroundContext);
+  if (!context) {
+    throw new Error("useBackground must be used within BackgroundProvider");
+  }
+  return context;
+}
+```
+
+### Usage
+```tsx
+// In layout or page
+<BackgroundProvider>
+  <YourComponents />
+</BackgroundProvider>
+
+// In components
+const { selectedYear, setSelectedYear } = useBackground();
+```
