--- src/components/custom/ContentArea/content-area.constants.ts
@@ -1,4 +1,0 @@
-/**
- * Shared constants for ContentArea component
- */
-export const CONTENT_VIEW_HEIGHT = 230;
