--- src/components/custom/Timeline/timeline.constants.ts
@@ -1,2 +1,0 @@
-export const YEAR_BUTTON_HEIGHT = 36;
-export const ERROR_BLINK_DURATION = 300;
